All tests passed

+++ WebTask/Models/OrderDetails.cs
@@ -9,7 +9,7 @@ namespace WebTask.Models
         public int OrderID { get; set; }
         public int ProductID { get; set; }
         public decimal UnitPrice { get; set; }
-        public int Quantity { get; set; }
-        public double Discount { get; set; }
+        public short Quantity { get; set; }
+        public float Discount { get; set; }
     }
 }
